pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "The backend of Food Delivery Web Application"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django[bcrypt]>=5.2.6",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.1",
    "django-phonenumber-field[phonenumbers]>=8.1.0",
    "django-scalar>=0.2.0",
    "django-split-settings>=1.3.2",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.28.0",
    "logtail-python>=0.3.4",
    "mongoengine>=0.29.1",
    "psycopg2-binary>=2.9.10",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "django-stubs>=5.2.5",
    "flake8>=7.3.0",
    "isort>=6.0.1",
    "mypy>=1.18.2",
    "pre-commit>=4.3.0",
]
